embed -dynamic interface into -compiled to support ts declare and use an unify Details class
